+++ boms/geode-all-bom/src/test/resources/expected-pom.xml
@@ -694,6 +694,11 @@
       </dependency>
       <dependency>
         <groupId>org.apache.geode</groupId>
+        <artifactId>geode-old-admin</artifactId>
+        <version>1.10.0-SNAPSHOT</version>
+      </dependency>
+      <dependency>
+        <groupId>org.apache.geode</groupId>
         <artifactId>geode-old-client-support</artifactId>
         <version>1.10.0-SNAPSHOT</version>
       </dependency>
