+++ geode-core/src/main/java/org/apache/geode/internal/DSFIDFactory.java
@@ -24,8 +24,6 @@ import it.unimi.dsi.fastutil.ints.Int2ObjectOpenHashMap;
 
 import org.apache.geode.DataSerializer;
 import org.apache.geode.InternalGemFireError;
-import org.apache.geode.admin.internal.SystemMemberCacheEventProcessor;
-import org.apache.geode.admin.jmx.internal.StatAlertNotification;
 import org.apache.geode.annotations.Immutable;
 import org.apache.geode.cache.InterestResultPolicy;
 import org.apache.geode.cache.client.internal.CacheServerLoadMessage;
@@ -103,95 +101,17 @@ import org.apache.geode.distributed.internal.membership.gms.messages.RemoveMembe
 import org.apache.geode.distributed.internal.membership.gms.messages.SuspectMembersMessage;
 import org.apache.geode.distributed.internal.membership.gms.messages.ViewAckMessage;
 import org.apache.geode.distributed.internal.streaming.StreamingOperation.StreamingReplyMessage;
-import org.apache.geode.internal.admin.ClientMembershipMessage;
-import org.apache.geode.internal.admin.remote.AddHealthListenerRequest;
-import org.apache.geode.internal.admin.remote.AddHealthListenerResponse;
-import org.apache.geode.internal.admin.remote.AddStatListenerRequest;
-import org.apache.geode.internal.admin.remote.AddStatListenerResponse;
-import org.apache.geode.internal.admin.remote.AdminConsoleDisconnectMessage;
-import org.apache.geode.internal.admin.remote.AdminConsoleMessage;
 import org.apache.geode.internal.admin.remote.AdminFailureResponse;
 import org.apache.geode.internal.admin.remote.AlertLevelChangeMessage;
 import org.apache.geode.internal.admin.remote.AlertListenerMessage;
-import org.apache.geode.internal.admin.remote.AlertsNotificationMessage;
-import org.apache.geode.internal.admin.remote.AppCacheSnapshotMessage;
-import org.apache.geode.internal.admin.remote.BridgeServerRequest;
-import org.apache.geode.internal.admin.remote.BridgeServerResponse;
-import org.apache.geode.internal.admin.remote.CacheConfigRequest;
-import org.apache.geode.internal.admin.remote.CacheConfigResponse;
-import org.apache.geode.internal.admin.remote.CacheInfoRequest;
-import org.apache.geode.internal.admin.remote.CacheInfoResponse;
-import org.apache.geode.internal.admin.remote.CancelStatListenerRequest;
-import org.apache.geode.internal.admin.remote.CancelStatListenerResponse;
-import org.apache.geode.internal.admin.remote.CancellationMessage;
-import org.apache.geode.internal.admin.remote.ChangeRefreshIntervalMessage;
 import org.apache.geode.internal.admin.remote.ClientHealthStats;
-import org.apache.geode.internal.admin.remote.CompactRequest;
-import org.apache.geode.internal.admin.remote.CompactResponse;
-import org.apache.geode.internal.admin.remote.DestroyEntryMessage;
-import org.apache.geode.internal.admin.remote.DestroyRegionMessage;
-import org.apache.geode.internal.admin.remote.DurableClientInfoRequest;
-import org.apache.geode.internal.admin.remote.DurableClientInfoResponse;
-import org.apache.geode.internal.admin.remote.FetchDistLockInfoRequest;
-import org.apache.geode.internal.admin.remote.FetchDistLockInfoResponse;
-import org.apache.geode.internal.admin.remote.FetchHealthDiagnosisRequest;
-import org.apache.geode.internal.admin.remote.FetchHealthDiagnosisResponse;
-import org.apache.geode.internal.admin.remote.FetchHostRequest;
-import org.apache.geode.internal.admin.remote.FetchHostResponse;
-import org.apache.geode.internal.admin.remote.FetchResourceAttributesRequest;
-import org.apache.geode.internal.admin.remote.FetchResourceAttributesResponse;
-import org.apache.geode.internal.admin.remote.FetchStatsRequest;
-import org.apache.geode.internal.admin.remote.FetchStatsResponse;
-import org.apache.geode.internal.admin.remote.FetchSysCfgRequest;
-import org.apache.geode.internal.admin.remote.FetchSysCfgResponse;
-import org.apache.geode.internal.admin.remote.FlushAppCacheSnapshotMessage;
-import org.apache.geode.internal.admin.remote.HealthListenerMessage;
-import org.apache.geode.internal.admin.remote.LicenseInfoRequest;
-import org.apache.geode.internal.admin.remote.LicenseInfoResponse;
 import org.apache.geode.internal.admin.remote.MissingPersistentIDsRequest;
 import org.apache.geode.internal.admin.remote.MissingPersistentIDsResponse;
-import org.apache.geode.internal.admin.remote.ObjectDetailsRequest;
-import org.apache.geode.internal.admin.remote.ObjectDetailsResponse;
-import org.apache.geode.internal.admin.remote.ObjectNamesRequest;
-import org.apache.geode.internal.admin.remote.ObjectNamesResponse;
 import org.apache.geode.internal.admin.remote.PrepareRevokePersistentIDRequest;
-import org.apache.geode.internal.admin.remote.RefreshMemberSnapshotRequest;
-import org.apache.geode.internal.admin.remote.RefreshMemberSnapshotResponse;
-import org.apache.geode.internal.admin.remote.RegionAttributesRequest;
-import org.apache.geode.internal.admin.remote.RegionAttributesResponse;
-import org.apache.geode.internal.admin.remote.RegionRequest;
-import org.apache.geode.internal.admin.remote.RegionResponse;
-import org.apache.geode.internal.admin.remote.RegionSizeRequest;
-import org.apache.geode.internal.admin.remote.RegionSizeResponse;
-import org.apache.geode.internal.admin.remote.RegionStatisticsRequest;
-import org.apache.geode.internal.admin.remote.RegionStatisticsResponse;
-import org.apache.geode.internal.admin.remote.RegionSubRegionSizeRequest;
-import org.apache.geode.internal.admin.remote.RegionSubRegionsSizeResponse;
-import org.apache.geode.internal.admin.remote.RemoveHealthListenerRequest;
-import org.apache.geode.internal.admin.remote.RemoveHealthListenerResponse;
-import org.apache.geode.internal.admin.remote.ResetHealthStatusRequest;
-import org.apache.geode.internal.admin.remote.ResetHealthStatusResponse;
 import org.apache.geode.internal.admin.remote.RevokePersistentIDRequest;
 import org.apache.geode.internal.admin.remote.RevokePersistentIDResponse;
-import org.apache.geode.internal.admin.remote.RootRegionRequest;
-import org.apache.geode.internal.admin.remote.RootRegionResponse;
-import org.apache.geode.internal.admin.remote.ShutdownAllGatewayHubsRequest;
 import org.apache.geode.internal.admin.remote.ShutdownAllRequest;
 import org.apache.geode.internal.admin.remote.ShutdownAllResponse;
-import org.apache.geode.internal.admin.remote.SnapshotResultMessage;
-import org.apache.geode.internal.admin.remote.StatAlertsManagerAssignMessage;
-import org.apache.geode.internal.admin.remote.StatListenerMessage;
-import org.apache.geode.internal.admin.remote.StoreSysCfgRequest;
-import org.apache.geode.internal.admin.remote.StoreSysCfgResponse;
-import org.apache.geode.internal.admin.remote.SubRegionRequest;
-import org.apache.geode.internal.admin.remote.SubRegionResponse;
-import org.apache.geode.internal.admin.remote.TailLogRequest;
-import org.apache.geode.internal.admin.remote.TailLogResponse;
-import org.apache.geode.internal.admin.remote.UpdateAlertDefinitionMessage;
-import org.apache.geode.internal.admin.remote.VersionInfoRequest;
-import org.apache.geode.internal.admin.remote.VersionInfoResponse;
-import org.apache.geode.internal.admin.statalerts.GaugeThresholdDecoratorImpl;
-import org.apache.geode.internal.admin.statalerts.NumberThresholdDecoratorImpl;
 import org.apache.geode.internal.cache.AddCacheServerProfileMessage;
 import org.apache.geode.internal.cache.BucketAdvisor;
 import org.apache.geode.internal.cache.CacheDistributionAdvisor;
@@ -527,8 +447,6 @@ public class DSFIDFactory implements DataSerializableFixedID {
     registerDSFID(DLOCK_REQUEST_MESSAGE, DLockRequestProcessor.DLockRequestMessage.class);
     registerDSFID(DLOCK_RESPONSE_MESSAGE, DLockRequestProcessor.DLockResponseMessage.class);
     registerDSFID(DLOCK_RELEASE_MESSAGE, DLockReleaseProcessor.DLockReleaseMessage.class);
-    registerDSFID(ADMIN_CACHE_EVENT_MESSAGE,
-        SystemMemberCacheEventProcessor.SystemMemberCacheMessage.class);
     registerDSFID(CQ_ENTRY_EVENT, CqEntry.class);
     registerDSFID(REQUEST_IMAGE_MESSAGE, InitialImageOperation.RequestImageMessage.class);
     registerDSFID(IMAGE_REPLY_MESSAGE, InitialImageOperation.ImageReplyMessage.class);
@@ -710,74 +628,13 @@ public class DSFIDFactory implements DataSerializableFixedID {
     registerDSFID(PR_REMOVE_BUCKET_REPLY, RemoveBucketReplyMessage.class);
     registerDSFID(PR_MOVE_BUCKET_MESSAGE, MoveBucketMessage.class);
     registerDSFID(PR_MOVE_BUCKET_REPLY, MoveBucketReplyMessage.class);
-    registerDSFID(ADD_HEALTH_LISTENER_REQUEST, AddHealthListenerRequest.class);
-    registerDSFID(ADD_HEALTH_LISTENER_RESPONSE, AddHealthListenerResponse.class);
-    registerDSFID(ADD_STAT_LISTENER_REQUEST, AddStatListenerRequest.class);
-    registerDSFID(ADD_STAT_LISTENER_RESPONSE, AddStatListenerResponse.class);
-    registerDSFID(ADMIN_CONSOLE_DISCONNECT_MESSAGE, AdminConsoleDisconnectMessage.class);
-    registerDSFID(ADMIN_CONSOLE_MESSAGE, AdminConsoleMessage.class);
+
     registerDSFID(MANAGER_STARTUP_MESSAGE, ManagerStartupMessage.class);
     registerDSFID(JMX_MANAGER_LOCATOR_REQUEST, JmxManagerLocatorRequest.class);
     registerDSFID(JMX_MANAGER_LOCATOR_RESPONSE, JmxManagerLocatorResponse.class);
     registerDSFID(ADMIN_FAILURE_RESPONSE, AdminFailureResponse.class);
     registerDSFID(ALERT_LEVEL_CHANGE_MESSAGE, AlertLevelChangeMessage.class);
     registerDSFID(ALERT_LISTENER_MESSAGE, AlertListenerMessage.class);
-    registerDSFID(APP_CACHE_SNAPSHOT_MESSAGE, AppCacheSnapshotMessage.class);
-    registerDSFID(BRIDGE_SERVER_REQUEST, BridgeServerRequest.class);
-    registerDSFID(BRIDGE_SERVER_RESPONSE, BridgeServerResponse.class);
-    registerDSFID(CACHE_CONFIG_REQUEST, CacheConfigRequest.class);
-    registerDSFID(CACHE_CONFIG_RESPONSE, CacheConfigResponse.class);
-    registerDSFID(CACHE_INFO_REQUEST, CacheInfoRequest.class);
-    registerDSFID(CACHE_INFO_RESPONSE, CacheInfoResponse.class);
-    registerDSFID(CANCELLATION_MESSAGE, CancellationMessage.class);
-    registerDSFID(CANCEL_STAT_LISTENER_REQUEST, CancelStatListenerRequest.class);
-    registerDSFID(CANCEL_STAT_LISTENER_RESPONSE, CancelStatListenerResponse.class);
-    registerDSFID(DESTROY_ENTRY_MESSAGE, DestroyEntryMessage.class);
-    registerDSFID(ADMIN_DESTROY_REGION_MESSAGE, DestroyRegionMessage.class);
-    registerDSFID(FETCH_DIST_LOCK_INFO_REQUEST, FetchDistLockInfoRequest.class);
-    registerDSFID(FETCH_DIST_LOCK_INFO_RESPONSE, FetchDistLockInfoResponse.class);
-    registerDSFID(FETCH_HEALTH_DIAGNOSIS_REQUEST, FetchHealthDiagnosisRequest.class);
-    registerDSFID(FETCH_HEALTH_DIAGNOSIS_RESPONSE, FetchHealthDiagnosisResponse.class);
-    registerDSFID(FETCH_HOST_REQUEST, FetchHostRequest.class);
-    registerDSFID(FETCH_HOST_RESPONSE, FetchHostResponse.class);
-    registerDSFID(FETCH_RESOURCE_ATTRIBUTES_REQUEST, FetchResourceAttributesRequest.class);
-    registerDSFID(FETCH_RESOURCE_ATTRIBUTES_RESPONSE, FetchResourceAttributesResponse.class);
-    registerDSFID(FETCH_STATS_REQUEST, FetchStatsRequest.class);
-    registerDSFID(FETCH_STATS_RESPONSE, FetchStatsResponse.class);
-    registerDSFID(FETCH_SYS_CFG_REQUEST, FetchSysCfgRequest.class);
-    registerDSFID(FETCH_SYS_CFG_RESPONSE, FetchSysCfgResponse.class);
-    registerDSFID(FLUSH_APP_CACHE_SNAPSHOT_MESSAGE, FlushAppCacheSnapshotMessage.class);
-    registerDSFID(HEALTH_LISTENER_MESSAGE, HealthListenerMessage.class);
-    registerDSFID(OBJECT_DETAILS_REQUEST, ObjectDetailsRequest.class);
-    registerDSFID(OBJECT_DETAILS_RESPONSE, ObjectDetailsResponse.class);
-    registerDSFID(OBJECT_NAMES_REQUEST, ObjectNamesRequest.class);
-    registerDSFID(LICENSE_INFO_REQUEST, LicenseInfoRequest.class);
-    registerDSFID(LICENSE_INFO_RESPONSE, LicenseInfoResponse.class);
-    registerDSFID(OBJECT_NAMES_RESPONSE, ObjectNamesResponse.class);
-    registerDSFID(REGION_ATTRIBUTES_REQUEST, RegionAttributesRequest.class);
-    registerDSFID(REGION_ATTRIBUTES_RESPONSE, RegionAttributesResponse.class);
-    registerDSFID(REGION_REQUEST, RegionRequest.class);
-    registerDSFID(REGION_RESPONSE, RegionResponse.class);
-    registerDSFID(REGION_SIZE_REQUEST, RegionSizeRequest.class);
-    registerDSFID(REGION_SIZE_RESPONSE, RegionSizeResponse.class);
-    registerDSFID(REGION_STATISTICS_REQUEST, RegionStatisticsRequest.class);
-    registerDSFID(REGION_STATISTICS_RESPONSE, RegionStatisticsResponse.class);
-    registerDSFID(REMOVE_HEALTH_LISTENER_REQUEST, RemoveHealthListenerRequest.class);
-    registerDSFID(REMOVE_HEALTH_LISTENER_RESPONSE, RemoveHealthListenerResponse.class);
-    registerDSFID(RESET_HEALTH_STATUS_REQUEST, ResetHealthStatusRequest.class);
-    registerDSFID(RESET_HEALTH_STATUS_RESPONSE, ResetHealthStatusResponse.class);
-    registerDSFID(ROOT_REGION_REQUEST, RootRegionRequest.class);
-    registerDSFID(ROOT_REGION_RESPONSE, RootRegionResponse.class);
-    registerDSFID(SNAPSHOT_RESULT_MESSAGE, SnapshotResultMessage.class);
-    registerDSFID(STAT_LISTENER_MESSAGE, StatListenerMessage.class);
-    registerDSFID(STORE_SYS_CFG_REQUEST, StoreSysCfgRequest.class);
-    registerDSFID(STORE_SYS_CFG_RESPONSE, StoreSysCfgResponse.class);
-    registerDSFID(SUB_REGION_REQUEST, SubRegionRequest.class);
-    registerDSFID(SUB_REGION_RESPONSE, SubRegionResponse.class);
-    registerDSFID(TAIL_LOG_REQUEST, TailLogRequest.class);
-    registerDSFID(TAIL_LOG_RESPONSE, TailLogResponse.class);
-    registerDSFID(VERSION_INFO_REQUEST, VersionInfoRequest.class);
-    registerDSFID(VERSION_INFO_RESPONSE, VersionInfoResponse.class);
     registerDSFID(HIGH_PRIORITY_ACKED_MESSAGE, HighPriorityAckedMessage.class);
     registerDSFID(SERIAL_ACKED_MESSAGE, SerialAckedMessage.class);
     registerDSFID(BUCKET_PROFILE, BucketAdvisor.BucketProfile.class);
@@ -806,14 +663,7 @@ public class DSFIDFactory implements DataSerializableFixedID {
     registerDSFID(INTEREST_EVENT_MESSAGE, InterestEventMessage.class);
     registerDSFID(INTEREST_EVENT_REPLY_MESSAGE, InterestEventReplyMessage.class);
     registerDSFID(HA_EVENT_WRAPPER, HAEventWrapper.class);
-    registerDSFID(STAT_ALERTS_MGR_ASSIGN_MESSAGE, StatAlertsManagerAssignMessage.class);
-    registerDSFID(UPDATE_ALERTS_DEFN_MESSAGE, UpdateAlertDefinitionMessage.class);
-    registerDSFID(REFRESH_MEMBER_SNAP_REQUEST, RefreshMemberSnapshotRequest.class);
-    registerDSFID(REFRESH_MEMBER_SNAP_RESPONSE, RefreshMemberSnapshotResponse.class);
-    registerDSFID(REGION_SUB_SIZE_REQUEST, RegionSubRegionSizeRequest.class);
-    registerDSFID(REGION_SUB_SIZE_RESPONSE, RegionSubRegionsSizeResponse.class);
-    registerDSFID(CHANGE_REFRESH_INT_MESSAGE, ChangeRefreshIntervalMessage.class);
-    registerDSFID(ALERTS_NOTIF_MESSAGE, AlertsNotificationMessage.class);
+
     registerDSFID(FIND_DURABLE_QUEUE, FindDurableQueueMessage.class);
     registerDSFID(FIND_DURABLE_QUEUE_REPLY, FindDurableQueueReply.class);
     registerDSFID(CACHE_SERVER_LOAD_MESSAGE, CacheServerLoadMessage.class);
@@ -835,14 +685,9 @@ public class DSFIDFactory implements DataSerializableFixedID {
     registerDSFID(INITIAL_IMAGE_VERSIONED_OBJECT_LIST, InitialImageVersionedEntryList.class);
     registerDSFID(FIND_VERSION_TAG, FindVersionTagMessage.class);
     registerDSFID(VERSION_TAG_REPLY, VersionTagReply.class);
-    registerDSFID(DURABLE_CLIENT_INFO_REQUEST, DurableClientInfoRequest.class);
-    registerDSFID(DURABLE_CLIENT_INFO_RESPONSE, DurableClientInfoResponse.class);
     registerDSFID(CLIENT_INTEREST_MESSAGE, ClientInterestMessageImpl.class);
     registerDSFID(LATEST_LAST_ACCESS_TIME_MESSAGE, LatestLastAccessTimeMessage.class);
-    registerDSFID(STAT_ALERT_DEFN_NUM_THRESHOLD, NumberThresholdDecoratorImpl.class);
-    registerDSFID(STAT_ALERT_DEFN_GAUGE_THRESHOLD, GaugeThresholdDecoratorImpl.class);
     registerDSFID(CLIENT_HEALTH_STATS, ClientHealthStats.class);
-    registerDSFID(STAT_ALERT_NOTIFICATION, StatAlertNotification.class);
     registerDSFID(FILTER_INFO_MESSAGE, InitialImageOperation.FilterInfoMessage.class);
     registerDSFID(SIZED_BASED_LOAD_PROBE, SizedBasedLoadProbe.class);
     registerDSFID(PR_MANAGE_BACKUP_BUCKET_MESSAGE, ManageBackupBucketMessage.class);
@@ -877,7 +722,6 @@ public class DSFIDFactory implements DataSerializableFixedID {
     registerDSFID(PERSISTENT_MEMBERSHIP_FLUSH_REQUEST, MembershipFlushRequest.class);
     registerDSFID(SHUTDOWN_ALL_REQUEST, ShutdownAllRequest.class);
     registerDSFID(SHUTDOWN_ALL_RESPONSE, ShutdownAllResponse.class);
-    registerDSFID(CLIENT_MEMBERSHIP_MESSAGE, ClientMembershipMessage.class);
     registerDSFID(END_BUCKET_CREATION_MESSAGE, EndBucketCreationMessage.class);
     registerDSFID(PREPARE_BACKUP_REQUEST, PrepareBackupRequest.class);
     registerDSFID(BACKUP_RESPONSE, BackupResponse.class); // in older versions this was
@@ -887,8 +731,6 @@ public class DSFIDFactory implements DataSerializableFixedID {
     registerDSFID(FINISH_BACKUP_RESPONSE, BackupResponse.class); // for backwards compatibility map
                                                                  // FINISH_BACKUP_RESPONSE to
                                                                  // BackupResponse
-    registerDSFID(COMPACT_REQUEST, CompactRequest.class);
-    registerDSFID(COMPACT_RESPONSE, CompactResponse.class);
     registerDSFID(FLOW_CONTROL_PERMIT_MESSAGE, FlowControlPermitMessage.class);
     registerDSFID(REQUEST_FILTERINFO_MESSAGE, InitialImageOperation.RequestFilterInfoMessage.class);
     registerDSFID(PARALLEL_QUEUE_REMOVAL_MESSAGE, ParallelQueueRemovalMessage.class);
@@ -937,7 +779,6 @@ public class DSFIDFactory implements DataSerializableFixedID {
     registerDSFID(DIST_TX_THIN_ENTRY_STATE, TXEntryState.DistTxThinEntryState.class);
     registerDSFID(SERVER_PING_MESSAGE, ServerPingMessage.class);
     registerDSFID(PR_DESTROY_ON_DATA_STORE_MESSAGE, DestroyRegionOnDataStoreMessage.class);
-    registerDSFID(SHUTDOWN_ALL_GATEWAYHUBS_REQUEST, ShutdownAllGatewayHubsRequest.class);
     registerDSFID(BUCKET_COUNT_LOAD_PROBE, BucketCountLoadProbe.class);
     registerDSFID(GATEWAY_SENDER_QUEUE_ENTRY_SYNCHRONIZATION_MESSAGE,
         GatewaySenderQueueEntrySynchronizationOperation.GatewaySenderQueueEntrySynchronizationMessage.class);
