+++ geode-old-admin/build.gradle
@@ -1,0 +1,43 @@
+/*
+ * Licensed to the Apache Software Foundation (ASF) under one or more
+ * contributor license agreements.  See the NOTICE file distributed with
+ * this work for additional information regarding copyright ownership.
+ * The ASF licenses this file to You under the Apache License, Version 2.0
+ * (the "License"); you may not use this file except in compliance with
+ * the License.  You may obtain a copy of the License at
+ *
+ *      http://www.apache.org/licenses/LICENSE-2.0
+ *
+ * Unless required by applicable law or agreed to in writing, software
+ * distributed under the License is distributed on an "AS IS" BASIS,
+ * WITHOUT WARRANTIES OR CONDITIONS OF ANY KIND, either express or implied.
+ * See the License for the specific language governing permissions and
+ * limitations under the License.
+ */
+
+apply from: "${rootDir}/${scriptDir}/standard-subproject-configuration.gradle"
+apply from: "${project.projectDir}/../gradle/publish.gradle"
+
+dependencies {
+  api(platform(project(':boms:geode-all-bom')))
+  implementation(project(':geode-core'))
+  implementation('org.apache.logging.log4j:log4j-api')
+
+  implementation('javax.mail:javax.mail-api')
+  implementation('commons-modeler:commons-modeler') {
+    exclude module: 'commons-logging-api'
+    exclude module: 'mx4j-jmx'
+    exclude module: 'xml-apis'
+  }
+  implementation('mx4j:mx4j')
+  implementation('mx4j:mx4j-remote')
+  implementation('mx4j:mx4j-tools')
+
+  testImplementation('org.mockito:mockito-core')
+  testImplementation(project(':geode-junit'))
+
+  integrationTestImplementation(project(':geode-junit'))
+  integrationTestImplementation(project(':geode-core'))
+
+  distributedTestImplementation(project(':geode-dunit'))
+}
