+++ geode-old-admin/src/main/java/org/apache/geode/admin/jmx/internal/GemFireHealthConfigJmxImpl.java
@@ -40,8 +40,6 @@ import org.apache.geode.admin.internal.GemFireHealthConfigImpl;
  *
  * @since GemFire 3.5
  */
-@edu.umd.cs.findbugs.annotations.SuppressWarnings(
-    justification = "This class is deprecated. Also, any further changes so close to the release is inadvisable.")
 public class GemFireHealthConfigJmxImpl
     implements GemFireHealthConfig, ManagedResource, java.io.Serializable {
