+++ geode-old-admin/src/main/java/org/apache/geode/internal/admin/remote/FetchHealthDiagnosisResponse.java
@@ -25,6 +25,7 @@ import org.apache.geode.admin.GemFireHealth;
 import org.apache.geode.distributed.internal.DistributionManager;
 import org.apache.geode.distributed.internal.HealthMonitor;
 import org.apache.geode.distributed.internal.membership.InternalDistributedMember;
+import org.apache.geode.internal.admin.GeodeAdminDistributedSystemService;
 
 /**
  * The response to fetching the health diagnosis.
@@ -44,7 +45,8 @@ public class FetchHealthDiagnosisResponse extends AdminResponse {
     FetchHealthDiagnosisResponse m = new FetchHealthDiagnosisResponse();
     m.setRecipient(recipient);
     {
-      HealthMonitor hm = dm.getHealthMonitor(recipient);
+      HealthMonitor hm = dm.getSystem().getService(GeodeAdminDistributedSystemService.class)
+          .getHealthMonitor(recipient);
       if (hm.getId() == id) {
         m.diagnosis = hm.getDiagnosis(healthCode);
       }
