+++ geode-old-admin/src/main/resources/META-INF/services/org.apache.geode.distributed.internal.DistributedSystemService
@@ -1,0 +1,1 @@
+org.apache.geode.internal.admin.GeodeAdminDistributedSystemService
