+++ geode-old-admin/src/main/resources/META-INF/services/org.apache.geode.internal.cache.CacheService
@@ -1,0 +1,1 @@
+org.apache.geode.internal.admin.GeodeAdminService
