+++ geode-old-admin/src/main/resources/org/apache/geode/internal/admin/sanctioned-geode-old-admin-serializables.txt
@@ -1,0 +1,26 @@
+org/apache/geode/admin/AdminException,true,879398950879472021
+org/apache/geode/admin/AdminXmlException,true,-6848726449157550169
+org/apache/geode/admin/CacheDoesNotExistException,true,-1639933911265729978
+org/apache/geode/admin/GemFireHealth$Health,true,3039539430412151801,healthString:java/lang/String
+org/apache/geode/admin/GemFireMemberStatus,true,3389997790525991310,_bindAddress:java/lang/String,_clientHealthStats:java/util/Map,_clientHostNames:java/util/Map,_clientQueueSizes:java/util/Map,_connectedClients:java/util/Set,_connectedIncomingGateways:java/util/Map,_connectedPeers:java/util/Set,_connectedServers:java/util/Set,_freeHeapSize:long,_gatewayHubStatus:java/lang/Object,_gatewayQueueSizes:java/util/Map,_hostAddress:java/net/InetAddress,_isClient:boolean,_isConnected:boolean,_isGatewayHub:boolean,_isLocator:boolean,_isPrimaryGatewayHub:boolean,_isServer:boolean,_locators:java/lang/String,_maximumHeapSize:long,_mcastAddress:java/net/InetAddress,_mcastPort:int,_memberId:java/io/Serializable,_memberName:java/lang/String,_outgoingGateways:java/util/Map,_regionStatuses:java/util/Map,_serverPort:int,_unconnectedServers:java/util/Set,upTime:long
+org/apache/geode/admin/SystemMemberType,true,3284366994485749302,ordinal:int
+org/apache/geode/admin/UnmodifiableConfigurationException,true,-7653547392992060646
+org/apache/geode/admin/internal/CacheHealthConfigImpl,false,maxEventQueueSize:long,maxLoadTime:long,maxNetSearchTime:long,minHitRatio:double
+org/apache/geode/admin/internal/GemFireHealthConfigImpl,true,-6797673296902808018,hostName:java/lang/String,interval:int
+org/apache/geode/admin/internal/MemberHealthConfigImpl,true,3966032573073580490,maxMessageQueueSize:long,maxReplyTimeouts:long,maxRetransmissionRatio:double,maxVMProcessSize:long
+org/apache/geode/admin/internal/StatisticImpl,true,3899296873901634399,internalStat:org/apache/geode/internal/admin/Stat
+org/apache/geode/admin/jmx/internal/AgentImpl$StartupException,true,6614145962199330348
+org/apache/geode/admin/jmx/internal/AgentLauncher$Status,true,-7758402454664266174,baseName:java/lang/String,exception:java/lang/Throwable,msg:java/lang/String,pid:int,state:int
+org/apache/geode/admin/jmx/internal/ConfigAttributeInfo,true,-1918437700841687078,config:org/apache/geode/admin/jmx/internal/ConfigurationParameterJmxImpl
+org/apache/geode/admin/jmx/internal/ConfigurationParameterJmxImpl,true,-7822171853906772375,deserialized:boolean
+org/apache/geode/admin/jmx/internal/ConnectionNotificationFilterImpl,true,1
+org/apache/geode/admin/jmx/internal/DynamicManagedBean,true,4051924500150228160
+org/apache/geode/admin/jmx/internal/GemFireHealthConfigJmxImpl,true,1482719647163239953,delegate:org/apache/geode/admin/GemFireHealthConfig,health:org/apache/geode/admin/GemFireHealth,mbeanName:java/lang/String,modelMBean:javax/management/modelmbean/ModelMBean,objectName:javax/management/ObjectName
+org/apache/geode/admin/jmx/internal/ManagedResourceType,true,3752874768667480449,ordinal:int
+org/apache/geode/admin/jmx/internal/RefreshNotificationType,true,4376763592395613794,ordinal:int
+org/apache/geode/admin/jmx/internal/StatisticAttributeInfo,true,28022387514935560,stat:org/apache/geode/admin/Statistic
+org/apache/geode/internal/SystemAdmin$CombinedResources,false
+org/apache/geode/internal/admin/CompoundEntrySnapshot,true,5776382582897895718,allUserAttributes:java/util/Set,allValues:java/util/Set,hitRatio:float,hitRatioSum:double,hitResponders:long,lastAccessTime:long,lastModifiedTime:long,name:java/lang/Object,numHits:long,numMisses:long
+org/apache/geode/internal/admin/CompoundRegionSnapshot,true,6295026394298398004,allCacheLoaders:java/util/Set,allCacheWriters:java/util/Set,allCapControllers:java/util/Set,allConcLevels:java/util/Set,allCustomIdle:java/util/HashSet,allCustomTtl:java/util/HashSet,allDataPolicies:java/util/Set,allEntryIdleTimeout:java/util/Set,allEntryTtl:java/util/Set,allInitialCaps:java/util/Set,allKeyConstraints:java/util/Set,allListeners:java/util/Set,allLoadFactors:java/util/Set,allRegionIdleTimeout:java/util/Set,allRegionTtl:java/util/Set,allScopes:java/util/Set,allStatsEnabled:java/util/Set,allUserAttributes:java/util/Set,allValueConstraints:java/util/Set,hitRatio:float,hitRatioSum:double,hitResponders:long,lastAccessTime:long,lastModifiedTime:long,name:java/lang/String,numHits:long,numMisses:long
+org/apache/geode/internal/admin/StatAlert,true,5725457607122449170,definitionId:int,time:java/util/Date,values:java/lang/Number[]
+org/apache/geode/internal/admin/remote/EntryValueNodeImpl,false,fields:org/apache/geode/internal/admin/remote/EntryValueNodeImpl[],name:java/lang/String,primitive:boolean,primitiveVal:java/lang/Object,type:java/lang/String
