+++ settings.gradle
@@ -61,6 +61,7 @@ include 'geode-concurrency-test'
 include 'boms:geode-client-bom'
 include 'boms:geode-all-bom'
 include 'static-analysis:pmd-rules'
+include 'geode-old-admin'
 
 ['1.0.0-incubating',
  '1.1.0',
